+++ src/components/AgreementFilters/AgreementFilters.test.js
@@ -76,6 +76,10 @@ describe('AgreementFilters', () => {
     await Accordion('Tags').exists();
   });
 
+  test('renders the Documents Accordion', async () => {
+    await Accordion('Documents').exists();
+  });
+
   test('renders the CustomPropertiesFilter component', () => {
     const { getByText } = renderComponent;
     expect(getByText('CustomPropertiesFilter')).toBeInTheDocument();